Battle skeleton
-Added ActiveSkills and PassiveSkills
-Added fields of skills to unit
-Added methods for invoking actions
-Added auxiliary fields to BattleUnitsStack
-Changed return value of .getBattleArmy() in BattleArmy to stack reference
-Changed values of Range to double

+++ src/gamengine/Range.java
@@ -1,10 +1,10 @@
 package gamengine;
 
 public class Range {
-    public int min;
-    public int max;
+    public double min;
+    public double max;
 
-    public Range(int min, int max) {
+    public Range(double min, double max) {
         this.min = min;
         this.max = max;
     }
